pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true

[project]
name = "saas-analytics"
version = "0.1.0"
description = "ServicePay Analytics - Public demo of a SaaS product analytics platform"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.84.0",
    "sqlalchemy==2.0.47",
    "aiosqlite==0.22.1",
    "aiomysql==0.3.2",
    "openai==2.24.0",
    "greenlet==3.3.2",
    "agno==2.5.5",
    "mcp==1.26.0",
    "httpx==0.28.1",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.2.1",
    "pymysql==1.1.2",
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "sse-starlette==2.2.1",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.3.0",
]

[dependency-groups]
dev = [
    "ruff==0.15.4",
    "black==26.1.0",
    "isort==8.0.1",
    "pre-commit==4.5.1",
]

[project.scripts]
saas-analytics = "core.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
skip-string-normalization = false
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["__pycache__", "*.pyc", ".venv"]
